Added bids
- Added 2 new columns in the listing database. One to store the current price and other to store the winner after the seller closes the listings.
- Bid forms is working now with error checking ( It checks if the new bid is higher than the starting price and the last bid )

I decided to add the current price column to optimize the speed loading any listing page. Instead of making a for loop with all the bids for that listing and search the bigger all the time, now that value is already stored in the listing table. It is a trade-off :)